Vaulting, Login, Signup, Logout
In this update I decided to scrap the entire project from before and start fresh again. Not really the kind of thing you enjoy hearing on a time crunch, but I wanted to build this in a modular fashion. This started with the login, signup, and logout functionalities. Signup required baseline username and password requirements. Protecting user information is crucial, therefore all passwords stored will first been encrypted then stored. When logging in, a hash comparison is done with the hash password stored for the user. If the passwords are correct, then and only then may the user access files under the protected folder. Before authorizing a user, if a user were to attempt to access a webpage that has yet to be granted to them, the website will redirect the user back to the login screen -- providing confidentiality and security. Logging out will provide the same experience. A user cannot enter unless they choose to log back in.

Each step required different technologies or methodologies to ensure the project remains whole. Frontend which include the HTML and CSS files. The frontend folder divides content between assets, protected, and public content. new users and logged out users will have access to public files and assets. logged in users will have access to protected files. Backend required '.env', node.js, mysql, bcrypt, and express in order to show the functionalities committed today. .env will store any information that github will not be seeing because we want to maintain security of this project. Node.js is the server that allows users to connect to the application, utilizing JavaScript. MySQL is the database system used to store user information. bcrypt is used for hashing the password before it's stored in the database. express is a node.js framework to assist in routing and hosting sessions for us today.